requirements.txt
python-telegram-bot==21.4
gspread==6.1.2
oauth2client==4.1.3